+++ semver-project-gradle-plugin/main/kotlin/com/javiersc/semver/project/gradle/plugin/valuesources/VersionValueSource.kt
@@ -8,9 +8,7 @@ import com.javiersc.semver.project.gradle.plugin.internal.checkVersionIsHigherOr
 import com.javiersc.semver.project.gradle.plugin.internal.commitsMaxCount
 import com.javiersc.semver.project.gradle.plugin.internal.git.GitCache
 import com.javiersc.semver.project.gradle.plugin.internal.git.GitRef
-import com.javiersc.semver.project.gradle.plugin.internal.git.currentBranch
 import com.javiersc.semver.project.gradle.plugin.internal.git.headCommit
-import com.javiersc.semver.project.gradle.plugin.internal.git.lastVersionTagInCurrentBranch
 import com.javiersc.semver.project.gradle.plugin.internal.projectTagPrefix
 import com.javiersc.semver.project.gradle.plugin.internal.scopeProperty
 import com.javiersc.semver.project.gradle.plugin.internal.stageProperty
@@ -19,7 +17,6 @@ import com.javiersc.semver.project.gradle.plugin.semverExtension
 import com.javiersc.semver.project.gradle.plugin.tasks.CreateSemverTagTask
 import com.javiersc.semver.project.gradle.plugin.tasks.PushSemverTagTask
 import java.io.File
-import org.eclipse.jgit.api.Git
 import org.gradle.api.Project
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
@@ -33,18 +30,24 @@ internal abstract class VersionValueSource : ValueSource<String, VersionValueSou
         with(parameters) {
             val isSamePrefix: Boolean = tagPrefixProperty.get() == projectTagPrefix.get()
 
-            fun cache() =
-                GitCache(
-                    gitDir = parameters.gitDir.get(),
-                    maxCount = parameters.commitsMaxCount,
-                )
+            var cache: GitCache? = null
+            fun cache(): GitCache {
+                if (cache == null) {
+                    cache =
+                        GitCache(
+                            gitDir = parameters.gitDir.get(),
+                            maxCount = parameters.commitsMaxCount,
+                        )
+                }
+                return cache!!
+            }
 
             val lastSemver: GradleVersion =
                 cache().lastVersionInCurrentBranch(projectTagPrefix.get())
-            val lastVersionInCurrentBranch =
+            val lastVersionInCurrentBranch: List<String> =
                 cache().versionsInCurrentBranch(projectTagPrefix.get()).map(GradleVersion::toString)
 
-            val lastVersionCommitInCurrentBranch =
+            val lastVersionCommitInCurrentBranch: String? =
                 cache().lastVersionCommitInCurrentBranch(projectTagPrefix.get())?.hash
 
             val version: String =
@@ -68,24 +71,18 @@ internal abstract class VersionValueSource : ValueSource<String, VersionValueSou
                 )
 
             val gradleVersion = GradleVersion(version)
-            val mappedGradleVersion: GradleVersion =
-                versionMapper.get().map(gradleVersion).let(GradleVersion::invoke)
-
-            val gitData: SemverExtension.GitData = cache().git.buildGitData(tagPrefixProperty)
-            val mappedVersionAndGit: String =
-                versionAndGitMapper.get().map(mappedGradleVersion, gitData)
+            val mappedVersion: String = versionMapper.get().map(gradleVersion)
 
             checkVersionIsHigherOrSame(
-                version = mappedVersionAndGit,
+                version = mappedVersion,
                 lastVersionInCurrentBranch = lastSemver,
             )
 
-            mappedVersionAndGit
+            mappedVersion
         }
 
     internal interface Params : ValueSourceParameters {
         val versionMapper: Property<SemverExtension.VersionMapper>
-        val versionAndGitMapper: Property<SemverExtension.VersionAndGitMapper>
         val gitDir: Property<File>
         val commitsMaxCount: Property<Int>
         val tagPrefixProperty: Property<String>
@@ -107,7 +104,6 @@ internal abstract class VersionValueSource : ValueSource<String, VersionValueSou
                     project.provider { semverExtension.gitDir.get().asFile }
 
                 parameters.versionMapper.set(semverExtension.versionMapper)
-                parameters.versionAndGitMapper.set(semverExtension.versionAndGitMapper)
                 parameters.gitDir.set(gitDir)
                 val commitsMaxCount: Int =
                     project.commitsMaxCount.orNull ?: semverExtension.commitsMaxCount.get()
@@ -127,53 +123,3 @@ private val Project.isCreatingSemverTag: Boolean
         gradle.startParameter.taskNames.any { taskName: String ->
             taskName == CreateSemverTagTask.NAME || taskName == PushSemverTagTask.NAME
         }
-
-private fun Git.buildGitData(tagPrefix: Property<String>): SemverExtension.GitData {
-    return SemverExtension.GitData(
-        tag =
-            lastVersionTagInCurrentBranch(tagPrefix.get())?.let { tag ->
-                SemverExtension.GitData.Tag(
-                    name = tag.name,
-                    refName = tag.refName,
-                    commit =
-                        SemverExtension.GitData.Commit(
-                            message = tag.commit.message,
-                            fullMessage = tag.commit.fullMessage,
-                            hash = tag.commit.hash,
-                        ),
-                )
-            },
-        commit =
-            SemverExtension.GitData.Commit(
-                message = headCommit.commit.message,
-                fullMessage = headCommit.commit.fullMessage,
-                hash = headCommit.commit.hash,
-            ),
-        branch =
-            SemverExtension.GitData.Branch(
-                name = currentBranch.name,
-                refName = currentBranch.refName,
-                commits =
-                    currentBranch.commits.map { commit ->
-                        SemverExtension.GitData.Commit(
-                            message = commit.message,
-                            fullMessage = commit.fullMessage,
-                            hash = commit.hash,
-                        )
-                    },
-                tags =
-                    currentBranch.tags.map { tag ->
-                        SemverExtension.GitData.Tag(
-                            name = tag.name,
-                            refName = tag.refName,
-                            commit =
-                                SemverExtension.GitData.Commit(
-                                    message = tag.commit.message,
-                                    fullMessage = tag.commit.fullMessage,
-                                    hash = tag.commit.hash,
-                                ),
-                        )
-                    },
-            ),
-    )
-}